requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.0
transformers==4.35.0
torch>=2.0.0
pydantic==2.4.2
python-multipart==0.0.6
nltk==3.8.1
rake-nltk==1.0.6
requests==2.31.0